requirements.txt
streamlit
numpy
opencv-python
matplotlib
scikit-image
Pillow